backend/requirements.txt
flask==2.3.3
flask-cors==4.0.1
torch==2.0.1
torchvision==0.15.2
pandas==2.1.0
pillow==10.0.0
openpyxl==3.1.2